% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{attach}
\alias{attach}
\alias{pyImport}
\alias{pyFunction}
\alias{pyAttach}
\title{Attach Python objects to R.}
\usage{
pyImport(name, from = NULL, as = NULL, env = parent.frame())

pyFunction(key, finalizer = "_")

pyAttach(what, env = parent.frame())
}
\arguments{
\item{name}{Names of the Python objects to import.}

\item{from}{(Optional) the name of the module.}

\item{as}{(Optional) An alias for the module name.}

\item{env}{(Optional) environment to assign the Python objects to.}

\item{key}{The Python function name.}

\item{finalizer}{An additional operation to perform on the
function output prior to returning to R. Use the character
"_" as a placeholder for the Python output variable.}

\item{what}{Vector of names of Python objects to attach to R.}
}
\description{
Attach Python objects to R.
}
\section{Functions}{
\itemize{
\item \code{pyImport}: Import Python objects into R.

\item \code{pyFunction}: import a Python function into R.

\item \code{pyAttach}: 
}}

