% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.r
\name{basics}
\alias{basics}
\alias{pyHelp}
\alias{pyInfo}
\alias{pyIsConnected}
\alias{pyExit}
\alias{pyConnect}
\title{Python Interface Basics}
\usage{
pyHelp(key)

pyInfo()

pyIsConnected()

pyExit()

pyConnect(path, port = 6000L, host = "localhost", timeout = 10000L)
}
\arguments{
\item{key}{The Python function or object to get help on.}

\item{path}{The path to the Python executable.}

\item{port}{A port number to use for communicating with Python.}

\item{host}{The hostname to use for communicating with Python.}

\item{timeout}{Maximum time to wait for a response from Python.}
}
\description{
Python Interface Basics
}
\section{Functions}{
\itemize{
\item \code{pyHelp}: Print Python help documents.

\item \code{pyInfo}: Print Python version information.

\item \code{pyIsConnected}: Check if Python is connected.

\item \code{pyExit}: Disconnect from Python.

\item \code{pyConnect}: Connect to Python.
}}

