% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{py_attach}
\alias{py_attach}
\alias{pyImport}
\alias{pyFunction}
\title{Attach Python objects to R.}
\usage{
pyImport(name, env = parent.frame())

pyFunction(key, finalizer = "_")
}
\arguments{
\item{name}{Name of the Python objemodule to import.}

\item{env}{(Optional) environment to assign the Python objects to.}

\item{key}{The Python function name.}

\item{finalizer}{An additional operation to perform on the
function output prior to returning to R. Use the character
"_" as a placeholder for the Python output variable.}
}
\description{
Attach Python objects to R.
}
\section{Functions}{
\itemize{
\item \code{pyImport}: Import Python module into R.

\item \code{pyFunction}: import a Python function into R.
}}

