% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.r
\name{py_interact}
\alias{py_interact}
\alias{pyExec}
\alias{pyExecp}
\alias{pyExecg}
\alias{pyExecfile}
\title{Interact with Python.}
\usage{
pyExec(code)

pyExecp(code)

pyExecg(code, returnValues = character())

pyExecfile(filename)
}
\arguments{
\item{code}{The Python code to execute.}

\item{returnValues}{A vector of variable names to
retrieve from Python.}

\item{filename}{The file containing Python code to execute.}
}
\value{
A list of variables from Python.
}
\description{
Interact with Python.
}
\section{Functions}{
\itemize{
\item \code{pyExec}: Execute a single line of Python code.

\item \code{pyExecp}: Execute a single line of Python code and
print the result.

\item \code{pyExecg}: Execute multiple line of Python code and
retrieve multiple variables using \code{pyGet}.

\item \code{pyExecfile}: Execute Python code contained in a file.
}}

