% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.r
\name{py_use}
\alias{py_use}
\alias{useNumpy}
\alias{usePandas}
\alias{useClass}
\title{Use Python Classes}
\usage{
useNumpy()

usePandas()

useClass(method)
}
\arguments{
\item{method}{A Python function definition for serializing the
class to a JSON string.}
}
\description{
Define methods for retrieving non-primitive Python objects.
}
\section{Functions}{
\itemize{
\item \code{useNumpy}: Use numpy classes.

\item \code{usePandas}: Use pandas classes.

\item \code{useClass}: Use a custom class.
}}

\seealso{
transpyr::use_class
}
