% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{variables}
\alias{variables}
\alias{pyPrint}
\alias{pyGet}
\alias{pySet}
\title{Interact with Python variables.}
\usage{
pyPrint(objName)

pyGet(...)

pySet(...)
}
\arguments{
\item{objName}{The Python object to print.}

\item{...}{Names of one or more Python variables to get/set.}
}
\value{
A character string.

The return value (for one variable) or a named list of
values (for multiple variables) returned from Python.
}
\description{
Interact with Python variables.
}
\details{
This is basically the same as \code{pyExecp} but prints to the
console using \code{print} instead of \code{cat}.
}
\section{Functions}{
\itemize{
\item \code{pyPrint}: Print a Python object to the R console.

\item \code{pyGet}: Retrieve variables from Python. Only applicable
to objects that can be serialized to JSON. Otherwise, use \code{pyPrint}.

\item \code{pySet}: Set variables in Python. Only applicable to
objects that can be serialized to JSON. Otherwise, use \code{pyExec}.
}}

