% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{attach}
\alias{attach}
\alias{pyImport}
\alias{pyFunction}
\alias{pyAttach}
\title{Attach Python objects to R.}
\usage{
pyImport(name, from = NULL, as = NULL, attach = parent.frame())

pyFunction(key)

pyAttach(what, env = parent.frame())
}
\arguments{
\item{from}{(Optional) the name of the module.}

\item{as}{(Optional) An alias for the module name.}

\item{key}{The Python function name.}

\item{what}{Vector of names of Python objects to attach to R.}

\item{env}{(Optional) environment to assign the Python objects to.}

\item{import}{Names of the Python objects to import.}
}
\description{
Attach Python objects to R.
}
\section{Functions}{
\itemize{
\item \code{pyImport}: Import Python objects into R.

\item \code{pyFunction}: import a Python function into R.

\item \code{pyAttach}: 
}}

